+++ src/input_block.rs
@@ -1,13 +1,13 @@
 use uuid::Uuid;
 
 pub struct InputBlock {
-    pub width: i8,
+    pub width: i32,
     pub pins: Vec<Uuid>,
 }
 
 impl InputBlock {
 
-    pub fn new(width: i8) -> Self {
+    pub fn new(width: i32) -> Self {
         let mut pins = Vec::new();
         for _i in 0..width {
             pins.push(Uuid::new_v4());
