+++ src/pin.rs
@@ -3,7 +3,7 @@ use uuid::Uuid;
 pub struct Pin {
     pub id: Uuid,
     pub neighbors: Vec<Uuid>,
-    pub state: i8,
+    pub state: i32,
 }
 
 impl Pin {
